Fixed jailkit scripts on fedora.
git-svn-id: svn://svn.ispconfig.org/ispconfig3/trunk@999 c1a6f2c7-8504-0410-933a-a4dae0184b0a

+++ server/plugins-available/shelluser_jailkit_plugin.inc.php
@@ -188,7 +188,7 @@ class shelluser_jailkit_plugin {
 				$tpl->setVar('home_dir',$this->_get_home_dir(""));
 				
 				$bashrc = escapeshellcmd($this->data['new']['dir']).'/etc/bash.bashrc';
-				exec('rm '.$bashrc);
+				if(@is_file($bashrc)) exec('rm '.$bashrc);
 				
 				file_put_contents($bashrc,$tpl->grab());
 				unset($tpl);
@@ -201,7 +201,7 @@ class shelluser_jailkit_plugin {
 				$tpl->setVar('domain',$web['domain']);
 				
 				$motd = escapeshellcmd($this->data['new']['dir']).'/var/run/motd';
-				exec('rm '.$motd);
+				if(@is_file($motd)) exec('rm '.$motd);
 				
 				file_put_contents($motd,$tpl->grab());
